+++ src/Pages/Proposal/Components/index.js
@@ -1,1 +1,2 @@
+export { default as ProposalDeposits } from './ProposalDeposits';
 export { default as ProposalInformation } from './ProposalInformation';